Fix parity bits for DID/SDID/count when generating 2038 VANC lines
The routine which generates a VANC line was not setting the parity
bits, which would in some cases cause checksum failures for the overall
payload.

A few notes:

Some 2038 TS streams (such as Adtec) will have the DID/SDID fields
populated with the parity bits present.  This is likely a violation
of the spec, but we handle it gracefully.  If parity bits are already
present, we don't touch them (simply passing them through).  This
prevents cases where if the parity in the original stream indicates
an error we don't recompute the parity and turn a byte which failed
the parity check into one that would now pass.

Problem manifested itself in the case where we were looking at TS
streams generated by our OBE implementation, where we don't set the
parity bits in the DID/SDID found in the PES.  Because our klvanc_capture
tool takes the PES content and generates a VANC line as an intermediate
step before invoking the parser, we would get checksum failures on
the resulting VANC line.  It's also likely that VANC lines we were
creating to  output to the Decklink module wouldn't be properly
setting the parity bits, but I suspect the driver and/or the FPGA
are computing parity for 10-bit streams if neither parity bit is
present in the input (hence we were feeding invalid data to the
output driver but the driver was quietly fixing it).