Add functions to manage generation of valid VANC lines from multiple packets
The SMPTE 291-1-2011 specification for SDI VANC has rules related to the
placement of VANC packets within VANC lines.  Add routines to enforce those
rules (in particular related to ensuring VANC packets are contiguous).  The
framework also allows for dealing with overlapping VANC packets (to prevent
some VANC packets from overwriting others).  This is particularly important
where want to place multiple VANC packets on a given line but we don't
specify/know/care where on the line they should appear.

Also ensure that we know go past the end-of-line with the set of VANC packets
intended for a given line.

Finally, have the function that generates the final VANC line ensure there
are no illegal sequences in the VANC packet which violate the spec.  If
such a malformed packet is found, it will no be included in the final line.