Fix search locations for includes/libraries
The existing tree had problems where it would prefer installed
versions of the headers/libraries over versions that were actually
inside the source tree.  Hence if there were a change to a function
prototype, the tree would fail to build because it would pick up
the header/libraries from /usr/local instead of the local copies.

Restructure the way we define the preprocessor and linker flags
to prioritize the local copies.  Changing CPPFLAGS to AM_CPPFLAGS
causes the local flags to appear earlier in the command line.  Also,
we now split out the options to pass to the preprocessor versus
the compiler (i.e. using CPPFLAGS versus CFLAGS where appropriate).

For the linker, we now just do an LDADD for the libtool reference
to the liklvanc library.  This eliminates the need to have the
-L../src/.libs argument, and also results in us properly detecting
and rebuilding tools if the library was changed (previously it
would indicate "Nothing to be done" even if the libklvanc.a changed).