Move splice request data to the actual MOM Op
Before this change, we had a single splice_request data structure
tied to the entire 104 packet.  However if the packet was a
Multiple Operation Message and there were multiple splice requests,
all of the operations would point to the same splice request data
(later Ops would trample earlier ops).  The net result is you would
have N sections created all with a copy of whatever was in the
last splice request.

Move the splice_request into the individual Op, so that you can
have multiple splice requests in a single Multiple Operation Message.

For now, disable all references to the splice data in Single
Operation Messages, which will eventually be removed entirely.