Remove padding code when generating VANC uint16_t array
When we first started working on the SCTE-104 output for Decklink,
we had a workaround for a bug in VLC which required the VANC
arrays to be padded on a 6-byte boundary.  We fixed that limitation
in VLC a while ago where we now don't have this requirement, and
in fact putting padding into the array like this will create gaps
between adjacent VANC packets in a given line (which would be a
violation of SMPTE ST 291-1-2011 Sec 7.3).

Remove the hack in libklvanc, now that we can rely on the calling
application doing the right thing.