Add functions to serialize MOM Ops into binary blobs
When parsing Multiple Operation Messages, we stash the original
binary data in an opaque 8-bit array in addition to any decoding
we do.  This works fine for cases where we parse and then
reserialize the data, but won't work if we need to change the
data and/or we are going from SCTE-35 to SCTE-104.  Create
functions which do the opposite of parsing the binary blob
into a struct, so we can create SCTE-104 message operations from
scratch.