Add new test tool which takes 104 packet, deserializes/re-serializes, then compares
This tool is specifically to exercise the serialization and
deserialization functionality related to SCTE-104 processing.  It
can take in pre-defined SCTE-104 VANC packets and look for issues
where we don't end up with matching data if we deserialize and
then re-serialize the data.

Only includes one sample packet for now, but it lays the
groundwork for a series of sample packets which can be part of
a regression test.