Don't do SCTE-104 over packet type 1 VANC messages
There is nothing in SMPTE 2010 that suggests that SCTE-104 can be
sent over type 1 VANC (and the one case where we saw it at a
customer it resulted in duplicated messages because the same
data was also being sent as type 2 packets).

Disable the functionality for now.  We can always re-enable it if
we find a case where it is necessary (or make it configurable).