Don't try to inject illegal value for tier in test messages
In the program that generates test SCTE-104 messages, I was setting
the tier to 0x1234, which is not a legal value (legal values are
between 0x0000 and 0x0fff).

We'll subsequently add a warning if we try to serialize out the
structure and the tier_data field contains an invalid value.