Pull later version of klbitstream_readwriter from libklscte35
There were apparently fixes related to writebits when passing it
a 64-bit int (as is common when serializing out PTS values).  Due
to the way the includes are setup within libklscte35 the copy from
libklvanc is being used instead.

We should look to make klbitstream local to libklvanc so it isn't
in the exported headers, but that can come as a later commit.  For
now just sync the copies up so that we address the 64-bit issue.