Add function to API to free packet_scte_104_s structures
Some of the existing code which uses libklvanc simply call free()
against the pointer, but this doesn't accommodate additional
complexity which can be found in the structure.  Add an actual API
call (which currently just calls free()), so that we have a single
routine which can handle any extra memory allocated.