Fix Coverity 1418199 (String not null terminated)
The code does an fread() call followed by printf()'ing the result,
but if a fully 80 bytes is read we will end up with a buffer that
is not null terminated.

Change the fread() call to ensure that the very last byte is always
a NULL character (the entire buffer is memset to zero immediately
prior to the fread call).