Fix regression that causes use-after-free
The following commit causes the pointer to the
memory being returned by the function call to
always be freed, even when there was no error
condition.

This results in a use-after-free on the first
SMPTE 2038 packet that gets decoded.

While it makes sense to do the free when hitting
the error case, we must avoid freeing the memory
on success, so we cannot simply drop through
to the error handling codepath.

commit 53deaa7650cf0a5e4c25331bcbb2d288a29c17a1
Author: Steven Toth <stoth@kernellabs.com>
Date:   Wed Mar 15 14:07:23 2017 -0400

    SCA: catch an edge case leak in the event of a validation failure and function exit.