Add a new parser that takes just the PES payload instead of the whole PES
FFmpeg already parses the PES headers and doesn't provide them to
downstream decoders, so we need a variant of the parser that doesn't
expect the header fields to be present.

Refactor parsing of the PES payload from the rest of the PES parsing,
and provide the ability for the user to call either function.