Don't invoke callbacks with malformed packets
If the VANC checksum is invalid, don't propagate the packet to
any callbacks.  This avoids the application having to be responsible
for checking the checksum field in every callback.

If the user is writing a debug tool and really wants the packets
to be announced anyway, he/she can set the allow_bad_checksums in
the VANC context.