Fix SCTE-104 test vectors to include proper packet length
When the original test vectors were generated, the messageSize field
was set to zero.  The tests passed until the bug was fixed, at which
point the actual (and correct) output from the parser no longer
matched the input test cases.

Fix those test cases such that the messageSize field is now properly
set, and update the checksum to match.