Don't report as success if VANC parser never ran.
Prior to this patch if the VANC parser choked on the checksum the
globally stored payload and its size from the previous result would
be used as the basis for comparison.  Change the code such that we
now explicitly set the resulting payload size to zero prior to feeding
the test vector into the parser.  This allows us to see that the
parser was never actually invoked and thus we can properly indicate
a failure and not show the output from the previous parsed packet.