Make parser call protocol-specific deallocation routines
This is basically the patch that was originally written by
Gang Liu <gang.liu@tfidm.com>, but tweaked slightly to deal
with the limitation that the internal parser expects the callbacks
to all have a (void *) argument.

This addresses memory leaks resulting from the assumption that all
of the packet types are a single flat allocation, which clearly
isn't true for SCTE-104.