Fix memory leak in conversion of SCTE-104 packets to VANC array
The routine which generates the final VANC array was reusing the
internal data member associated with each MOM type.  This is bad
because it will trample whatever might have already been in there,
as well as it leaking memory because it was blindly reallocating
the pointer.

It's also counter-intuitive that a routine which serializes a
struct into a byte array might actually change the contents
of the struct provided as input.

Use a temporary buffer to store the output bytes instead.  A
subsequent patch will mark the argument as const to make it clear
that it won't modify the struct.