Properly serialize VANC packets to V210 standard definition lines
While the HD-SDI specification calls for VANC packets to be written
to the Y region (i.e. skipping U/V fields), with standard definition
SDI both the Y and UV are expected to be used.

Provide a pixel conversion routine, as well as creating a new variant
of the klvanc_generate_vanc_line() function which takes care of
the serialization automatically (based on the line width passed as
an argument).  This means the caller no longer has to take any
responsibility for the final colorspace conversion and can just
pass in the final VANC line buffer and the pixel width.