Further break out AFD Bar Data into component parts
The original implementation just had stored the raw bytes, leaving
it to the application to do more extensive parsing.  Convert to
using explicitly enums and break out the individual fields into
members of the parsed packet.

Note this change also has a minor change to the dumping of
parameters where we no longer dump only the bottom 14 bits of the
16-bit word.  This is because of a mis-reading of the specs where
I assumed SMPTE 2016-3 was consistent with SCTE-128.  It turns up
SCTE-128 (encoding in H.264) only uses 14 bits and the two top
bits are marker bits, but in 2016-3 (encoding for SDI) the full
16-bits are used and the two high bits are zero.