Fix case where framerate provide isn't a reduced fraction
This isn't a typical case, but for non-NTSC framerates it's likely
that the framerate will not come as a fully reduced fraction
(e.g. 50000/1000 rather than 50/1).  In fact, the Blackmagic SDK
does this by default when you enumerate the supported video modes.

Add a routine to reduce the fraction to its LCD prior to doing the
lookup.