Don't treat AFD type 0x00 as invalid
According to SMPTE ST 2016-1 Sec 5 Note 1, it is valid to have an
AFD type of 0x00 (i.e. "Undefined").  In this context, "Undefined"
is not synonymous with an invalid value, and in fact is in some
cases used in conjunction with bar data.

See Note 1 under Table 1 of ST2016-1:2009 on Page 6 for the details
as to how to interpret AFD type 0000.

Remove the notion that AFD_UNDEFINED is the same as an invalid
value.  Because we synchronized the raw values with the values in
the enum in a previous commit, we no longer need to translate back
and forth between them.