Make klvanc_cache_reset() take much less time
Because of the core design of the cache, at initialization the
code allocates 255x255x2048 instances os klvanc_cache_line_s
(i.e. 133 million).  Iterating over these can take multiple seconds,
in particular with Clang which appears to not optimize the branch
prediction.

Add a quick check where if there were never any active entries found
on a DID/SDID, we don't bother iterating over the 2048 cache lines.

It's unusual to have more than a few different DID/SDIDs active
on any given SDI feed, so this reduces the cost to approx 1/65535.
On my laptop this results in the reset going from "waiting for
nearly ten seconds", to "virtually instantaneous".

For the record, the problem manifested as it taking nearly ten
seconds for klvanc_capture to exit from the time the user hit ctrl-c.