Fix pre-roll not being populated for SCTE-104 time signal
Fix the parser code to properly populate the pre_roll_time member
if we receive a SCTE-104 time_signal message.  We weren't setting
the value at all, which would result in it always being zero
when converting to SCTE-35.

Also add a test vector that includes a time_signal with non-zero
preroll.