Cleanup dependencies from tools being moved to klvanc-tools
The stuff that required zlib, ncurses, and libklmonitoring was all
moved to klvanc-tools, and are not required by the libklvanc library
itself.

Update the configure/Makefile.am to not require those components,
which makes the library more portable for those who want to use
the library without the klvanc-tools command-line utilities.