Make TC parsing conformant to SMPTE ST 334-2
There was an apparent bug in the CEA-708 B revision of the
specification where the intent was for the timecode section to be
five bytes but Table 26 added up to 41 bits (resulting in the rest
of the packet not being byte-aligned).  It is unclear whether any
implementation actually implemented this literally, but the
CEA acknowledged the issue in CEA 10-B and recommended treating
the struct as 40-bits, and changed the in subsequent versions
to reflect the change.

Change the parsing to reflect the corrected structure (which
was removed entirely in subsequent versions of CEA-708 and
is now formally defined in SMPTE ST 334-2).

References:

CEA-708-B Sec 11.2.3, Table 26
CEA 10-B 708-B Implementation Guidance Sec 30.9/30.10
SMPTE ST 334-2:2007, Sec 5.3, Table 4