Use both flags and section ID to determine whether to output section
When creating 708 packets, don't rely solely on the header field
to determine whether to create a section.  Also check for the
section ID.  This prevents invalid sections from being written
out.

This allows for consistency in cases where we deserialized a
packet that had the Service Info flag set but the section itself
wasn't actually present.