Fix error in math for short packet in svcinfo
The check for remaining bytes in the svcinfo section occurred
after reading the first byte (which might not actually exist).

Hence if the svc_count field was non-zero but there were no bytes
that follow, the read of the first eight bits would read past
the end of the buffer.

Perform the check before looking at any of the bytes, and move
the check outside the loop to improve readability.