Get rid of invalid entry in DID/SDID lookup table
We had an entry in the table that attempted to map DID/SDID
0x80/0x01 to the SCTE-104 parser.  We commented this out long ago
thinking that it was a type-1 packet, but in fact the packet is
marked as deleted.

Remove the dead reference entirely, since there is no reasonable
circumstance where we would want to feed a deleted packet into the
SCTE-104 parser.