Make sure test rig reports failure if callback not invoked
If the callback never fires, the contents of the global vancResult
contents from the previous run will be reused.  Make sure we reset
it prior to each run.

This also addresses the case where the parse call succeeds but the
resulting buffer length is zero (which should be treated as a
failure by the test harness).