Show proper description/specification for VANC types we can't decode
The routine the klvanc_capture tool uses to find the description
and specification strings to show made use of the pkt->type field
rather than the DID/SDID.  As a result, the curses GUI would show
UNKNOWN/UNKNOWN for any packet type which we don't have a decoder
for, even though we have a comprehensive list of all the registered
packet types from smpte-ra.org.

Use the version of the function which takes the DID/SDID, so that
we can show correct descriptions for packet types that we don't
explicitly have a decoder for.