Fix sorting for multiple VANC packets on a single line
If we are generating lines that have multiple VANC packets, the
sorting routine was broken due to incorrect setup of the comparison
function.  The result would be a non-deterministic ordering of
the packets.

Fix the comparison callback so we end up with consistent ordering
based on the h_offset field for the vanc entry.