Fix parsing for SMPTE ST 334-1 CEA-608 packets on DID/SDID 61/02
The original code assumed that the three byte payload contained in
the 61/02 was the same structure as a CEA-708 tuple.  However the
basic structure is different and the format for the first octet
is described in SMPTE ST 334-1:2015, Annex B.

Update the structure for the parsed packet to expose the correct
fields.