Improve 2038 parser robustness to properly handle data_count with parity
The data_count field is ten bits in size (including parity), and some
encoder implementations include the parity bits.  Because we store the
value in the l->data_count in the original form received, we need to make
sure in any cases we reference the value for determining the number of
user data words doesn't include the parity bits.

From SMPTE ST2038:2008 Sec 4.2.1 Paragraph 5:

"For the purposes of the “for” loop syntax counter only, only the
lower 8 bits of data_count, corresponding to bits b7 (MSB) through
 b0 (LSB) of the 10-bit SMPTE 291M data count word, shall be used."

For streams which contained the parity bits, we were attempting to
create an array which was much larger than the number of bytes in
the original packet, which was causing a subsequent safety check to
fail and the packets to be silently discarded as invalid.