Fix ctrl-c not working with file based sources
The test app is intercepting the SIGINT and setting a flag, but
the app will only terminate with a UDP source given the file
source never checks the flag in its main loop.

Make the main loop for file sources also check the flag, which is
important for cases where the TS file is very large and the user
wants to stop it before it runs to completion.