Fix bug in check for available bytes in ringbuffer detected by valgrind
The read operation expects to dequeue (pes_length + 2) bytes from the
ring buffer, but the check to make sure the ring buffer has enough
available data effectively checks for (pes_length - 2).  This can
result in the resulting PES packet containing uninitialized data at last
byte or two of the its buffer.

Reported by valgrind in a sample where the rb_used was 31 and the
pes_length was 30 (i.e. attempted to copy 32 bytes from the ring
buffer but only 31 was available).