Fix for properly padding chroma region of 10-bit VANC
Our conversion from Y10 to V210 was inserting 0x000 into the
chroma portion of the V210 output (which would get clamped to
0x004 by the decklink hardware).  The proper behavior is to set
the unused chroma values to 0x200.

While the actual VANC data is being stored in the luma region and
the chroma is unused, some downstream tools apparently expect to
find 0x200 in the chroma region, and fail to parse if otherwise.
So adjust our creation of the V210 region to properly pad the
chroma region with 0x200.  This patch also ensures any leftover
values in the luma range are set to 0x040.