Fix segfault with verbosity level 2 if callbacks not defined
A couple of years ago we refactored the code to not call the parser
on any VANC packets that don't have a callback defined.  This both
has a performance benefit (i.e. don't bother decoding packets the
user doesn't care about), as well as a safety benefit (i.e. bugs
in parsers won't crash the application if the user doesn't care
about those packets).

There is a logging routine though at verbosity level 2 which will
dump the decoded packet contents, but that didn't take into account
that the packet might have skipped decoding entirely.  This would
cause a segfault due to a NULL pointer dereference.

Skip the dump routine if the packet was not decoded.