Don't show ADF parsing outcome unless at the most verbose log level
The check for ADF is something that is very noisy, so should only
be enabled at the very highest log level.  This helps with cases
where the user turned up the logging to get some debug info but
doesn't want to see literally thousands of lines of extra logging
as the parser searches for the ADF header.