Add test case which crashes SCTE-104 parser
This was seen in the wild where we had a VANC packet with a valid
checksum, but the sixth MOM descriptor has a length of 29939
(thereby blowing past the end of the packet).

Add a test case that demonstrates this failure, which in the
current code causes a segfault on the memcpy() for the o->data
buffer.