Set parity bits properly for DID/SDID/DC in ST2038 packets generated
The code which creates the 2038 PES payload for the individual lines
will insert whatever DID/SD/DC values are provided by the user,
with no attempt to add the requisite parity bits if not provided.

Add the parity in as needed, so we conform to the specification
and work properly with downstream parsers which expect the parity
to be present.